[MonologBundle] Moved the handlers to the Bridge namespace
